# Conflicts:
#	client/src/main/java/com/alibaba/nacos/client/config/filter/impl/ConfigFilterChainManager.java
#	client/src/test/java/com/alibaba/nacos/client/config/filter/impl/ConfigFilterChainManagerTest.java

+++ naming/src/main/java/com/alibaba/nacos/naming/core/v2/metadata/NamingMetadataManager.java
@@ -76,7 +76,8 @@ public class NamingMetadataManager extends SmartSubscriber {
      * @return true if contain instance metadata, otherwise false
      */
     public boolean containInstanceMetadata(Service service, String metadataId) {
-        return instanceMetadataMap.containsKey(service) && instanceMetadataMap.get(service).containsKey(metadataId);
+        ConcurrentMap<String, InstanceMetadata> metadataMap = instanceMetadataMap.get(service);
+        return metadataMap != null && metadataMap.containsKey(metadataId);
     }
     
     /**
@@ -127,10 +128,7 @@ public class NamingMetadataManager extends SmartSubscriber {
      * @param instanceMetadata new instance metadata
      */
     public void updateInstanceMetadata(Service service, String metadataId, InstanceMetadata instanceMetadata) {
-        if (!instanceMetadataMap.containsKey(service)) {
-            instanceMetadataMap.putIfAbsent(service, new ConcurrentHashMap<>(INITIAL_CAPACITY));
-        }
-        instanceMetadataMap.get(service).put(metadataId, instanceMetadata);
+        instanceMetadataMap.computeIfAbsent(service, k -> new ConcurrentHashMap<>(INITIAL_CAPACITY)).put(metadataId, instanceMetadata);
     }
     
     /**
@@ -151,9 +149,11 @@ public class NamingMetadataManager extends SmartSubscriber {
      */
     public void removeInstanceMetadata(Service service, String metadataId) {
         ConcurrentMap<String, InstanceMetadata> instanceMetadataMapForService = instanceMetadataMap.get(service);
-        instanceMetadataMapForService.remove(metadataId);
-        if (instanceMetadataMapForService.isEmpty()) {
-            serviceMetadataMap.remove(service);
+        if (null != instanceMetadataMapForService) {
+            instanceMetadataMapForService.remove(metadataId);
+            if (instanceMetadataMapForService.isEmpty()) {
+                instanceMetadataMap.remove(service);
+            }
         }
         expiredMetadataInfos.remove(ExpiredMetadataInfo.newExpiredInstanceMetadata(service, metadataId));
     }
@@ -190,7 +190,8 @@ public class NamingMetadataManager extends SmartSubscriber {
      */
     public void loadServiceMetadataSnapshot(ConcurrentMap<Service, ServiceMetadata> snapshot) {
         for (Service each : snapshot.keySet()) {
-            ServiceManager.getInstance().getSingleton(each);
+            Service  service = Service.newService(each.getNamespace(), each.getGroup(), each.getName(), each.isEphemeral());
+            ServiceManager.getInstance().getSingleton(service);
         }
         ConcurrentMap<Service, ServiceMetadata> oldSnapshot = serviceMetadataMap;
         serviceMetadataMap = snapshot;
